quick POC of toNewPurchaseCallback
